loading element added
