+++ src/components/Loading.jsx
@@ -1,0 +1,6 @@
+const Loading = () => {
+  return (
+    <div>Loading...</div>
+  )
+}
+export default Loading
