+++ src/pages/user/SearchBookPage/components/SearchBookForm.jsx
@@ -2,19 +2,7 @@ import { Formik, Form, Field, ErrorMessage } from "formik"
 import * as Yup from "yup";
 import PropTypes from 'prop-types';
 
-const SearchBook = ({ setResults }) => {
-
-    const searchBook = async (author, title) => {
-        try {
-            const response = await fetch(`https://www.googleapis.com/books/v1/volumes?q=${title}+inauthor:${author}&key=${import.meta.env.VITE_GOOGLE_BOOKS_API_KEY}&maxResults=20`)
-            const data = await response.json()
-            console.log(data)
-            setResults(data)
-            console.log('dataitems', data.items)
-        } catch (err) {
-            console.log(err)
-        }
-    }
+const SearchBook = ({ searchBook }) => {
 
     const onSubmit = (values, { resetForm, setErrors }) => {
         if (!values.author && !values.title) {
@@ -73,5 +61,8 @@ const SearchBook = ({ setResults }) => {
 export default SearchBook
 
 SearchBook.propTypes = {
-    setResults: PropTypes.func
+    setResults: PropTypes.func,
+    setIsLoading: PropTypes.func,
+    searchBook: PropTypes.func,
+
 }
