+++ src/pages/user/SearchBookPage/components/SearchBookResults.jsx
@@ -10,11 +10,12 @@ const SearchBookResults = ({ results }) => {
             {results?.totalItems > 0 && booksResults(results.items)}
             {results?.totalItems === 0 && <p className="text-lg">Book not found.</p>}
         </>
+
     )
 }
 export default SearchBookResults
 
 SearchBookResults.propTypes = {
-    results: PropTypes.object, 
+    results: PropTypes.object,
     setResults: PropTypes.func
 }
